Change int Id to long Id

+++ EmpiriaBMS.Core/Dtos/DisciplineDto.cs
@@ -9,10 +9,10 @@ public class DisciplineDto : EntityDto
 
     public long EstimatedHours { get; set; }
 
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public DisciplineType Type { get; set; }
 
-    public int ProjectId { get; set; }
+    public long ProjectId { get; set; }
     public Project Project { get; set; }
 
     public float EstimatedCompleted { get; set; }
