+++ EmpiriaBMS.Core/Repositories/Base/IRepository.cs
@@ -7,19 +7,19 @@ public interface IRepository<T, U>
     where T : class, IEntityDto
     where U : class, IEntity
 {
-    Task<T?> Get(int id);
+    Task<T?> Get(long id);
     Task<ICollection<T>> GetAll(int pageSize = 0, int pageIndex = 0);
     Task<ICollection<T>> GetAll(
         Expression<Func<U, bool>> expresion,
         int pageSize = 0,
         int pageIndex = 0
     );
-    Task<int> Count();
-    Task<int> Count(Expression<Func<U, bool>> expresion);
+    Task<long> Count();
+    Task<long> Count(Expression<Func<U, bool>> expresion);
     Task<bool> Any(Expression<System.Func<U, bool>> expresion);
     Task<T> Add(T model, bool update);
     Task<T> Update(T model);
-    Task<T> Delete(int id);
+    Task<T> Delete(long id);
     Task SaveChangesAsync();
     void Dispose();
 }
