+++ EmpiriaBMS.Models/Models/Issue.cs
@@ -5,13 +5,13 @@ public class Issue : Entity
 {
     public DateTime ComplaintDate { get; set; }
 
-    public int ProjectId { get; set; }
+    public long ProjectId { get; set; }
     public Project? Project { get; set; }
 
-    public int DisplayedRoleId { get; set; }
+    public long DisplayedRoleId { get; set; }
     public Role? DisplayedRole { get; set; }
 
-    public int CreatorId { get; set; }
+    public long CreatorId { get; set; }
     public User? Creator { get; set; }
 
     public ICollection<Document>? Documents { get; set; }
